+++ __tests__/lib/constants/seo.test.ts
@@ -23,3 +23,4 @@ describe("SEO_CONFIG structured data", () => {
 
 
 
+