+++ supabase/migrations/20251224090000_add_location_ranking_coordinate_output_v2.sql
@@ -198,3 +198,4 @@ GRANT EXECUTE ON FUNCTION get_beaches_by_metro_with_scores_v2 TO anon;
 COMMIT;
 
 
+